requirements.txt
fastmcp>=2.11.0
python-dotenv>=1.0.0
pydantic>=2.11.0
uvicorn>=0.20.0
mcp>=1.12.0
